agentos/requirements.txt
# Base: autogenerated by uv, then extended with popular packages
# for agent development, data processing, and observability.
#
# --- Core framework ---
agno==2.5.10
fastapi==0.134.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.0
fastar==0.8.0
starlette==0.52.1
uvicorn==0.41.0
uvloop==0.22.1
mcp==1.26.0
#
# --- AI / LLM ---
openai==2.24.0
tiktoken==0.12.0
#
# --- Observability (Phoenix Arize + OpenTelemetry) ---
arize-phoenix-client==2.0.0
arize-phoenix-evals==2.11.0
openinference-instrumentation==0.1.45
openinference-instrumentation-agno==0.1.28
openinference-semantic-conventions==0.1.26
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
googleapis-common-protos==1.72.0
grpcio==1.71.0
protobuf==5.29.3
sentry-sdk==2.53.0
#
# --- Phoenix evals transitive deps ---
jsonpath-ng==1.8.0
pystache==0.6.8
#
# --- Data processing & analysis ---
numpy==2.4.2
pandas==3.0.1
pyarrow==23.0.1
tabulate==0.10.0
#
# --- Document parsing ---
beautifulsoup4==4.14.3
soupsieve==2.8.3
defusedxml==0.7.1
lxml==5.4.0
odfpy==1.4.1
openpyxl==3.1.5
pypdf==5.6.0
python-docx==1.1.2
#
# --- Image processing ---
Pillow==12.1.1
#
# --- Database ---
pgvector==0.4.2
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.10
sqlalchemy==2.0.47
#
# --- HTTP / networking ---
aiohttp==3.13.3
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
frozenlist==1.8.0
multidict==6.7.1
yarl==1.23.0
certifi==2026.2.25
charset-normalizer==3.4.5
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
requests==2.32.3
urllib3==2.6.3
websockets==16.0
#
# --- Pydantic / validation ---
annotated-doc==0.0.4
annotated-types==0.7.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
email-validator==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
#
# --- Resilience / retry ---
tenacity==9.1.4
#
# --- CLI / terminal ---
click==8.3.1
rich==14.3.3
rich-toolkit==0.19.7
shellingham==1.5.4
typer==0.24.1
pygments==2.19.2
tqdm==4.67.3
#
# --- Templating / markup ---
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
#
# --- Crypto / security ---
cffi==2.0.0
cryptography==46.0.5
pycparser==3.0
pyjwt==2.11.0
#
# --- Async / scheduling ---
anyio==4.12.1
croniter==6.0.0
sniffio==1.3.1
sse-starlette==3.3.2
#
# --- File watching ---
watchdog==6.0.0
watchfiles==1.1.1
#
# --- Stdlib extensions / misc ---
attrs==25.4.0
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
importlib-metadata==8.7.1
jiter==0.13.0
packaging==26.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
pyyaml==6.0.3
rignore==0.7.6
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
typing-extensions==4.15.0
typing-inspection==0.4.2
wrapt==1.17.3
zipp==3.23.0
